+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/example/user-config-resolver-go
 
-go 1.23.8
+go 1.23
 
 require github.com/Knetic/govaluate v3.0.0+incompatible
