+++ resolver/json/test_utils.go
@@ -1,4 +1,4 @@
-package jsonresolver
+package json
 
 import (
 	"encoding/json"
