+++ resolver/json/testdata/custom-user-groups/input.json
