+++ resolver/json/testdata/custom-user-groups/output.json
