+++ resolver/json/testdata/invalid-config/input.json
