+++ resolver/json/testdata/user-in-all-groups/input.json
