+++ resolver/json/testdata/user-in-all-groups/output.json
