+++ resolver/json/testdata/user-in-any-groups/input.json
