+++ resolver/json/testdata/user-in-any-groups/output.json
