+++ resolver/json/testdata/user-in-different-groups/input.json
