+++ resolver/json/testdata/user-in-different-groups/output.json
