+++ resolver/json/testdata/user-in-no-groups/input.json
