+++ resolver/json/testdata/user-in-no-groups/output.json
