+++ resolver/logic.go
